pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "f-admin-client"
version = "0.0.1"
dependencies = [
  "fastapi",
  "starlette",
  "tortoise-orm",
  "convert-case",
]
requires-python = ">=3.11"
readme = "README.md"
description = "1"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.10",
]